Change to localApplicationExecutor
