--- components/org.wso2.carbon.identity.application.authenticator.identifierauth/src/main/java/org.wso2.carbon.identity.application.authenticator.identifierauth/IdentifierAuthenticator.java
@@ -1,374 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.wso2.carbon.identity.application.authenticator.identifierauth;
-
-import com.hazelcast.util.Base64;
-import org.apache.commons.lang.StringUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.identity.application.authentication.framework.AbstractApplicationAuthenticator;
-import org.wso2.carbon.identity.application.authentication.framework.AuthenticatorFlowStatus;
-import org.wso2.carbon.identity.application.authentication.framework.LocalApplicationAuthenticator;
-import org.wso2.carbon.identity.application.authentication.framework.LocalApplicationExecutor;
-import org.wso2.carbon.identity.application.authentication.framework.config.ConfigurationFacade;
-import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
-import org.wso2.carbon.identity.application.authentication.framework.exception.AuthenticationFailedException;
-import org.wso2.carbon.identity.application.authentication.framework.exception.InvalidCredentialsException;
-import org.wso2.carbon.identity.application.authentication.framework.exception.LogoutFailedException;
-import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
-import org.wso2.carbon.identity.application.authentication.framework.util.FrameworkUtils;
-import org.wso2.carbon.identity.application.authenticator.identifierauth.internal
-        .IdentifierAuthenticatorServiceComponent;
-import org.wso2.carbon.identity.application.common.model.User;
-import org.wso2.carbon.identity.base.IdentityRuntimeException;
-import org.wso2.carbon.identity.core.model.IdentityErrorMsgContext;
-import org.wso2.carbon.identity.core.util.IdentityCoreConstants;
-import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
-import org.wso2.carbon.identity.core.util.IdentityUtil;
-import org.wso2.carbon.user.api.UserRealm;
-import org.wso2.carbon.user.core.UserCoreConstants;
-import org.wso2.carbon.user.core.UserStoreException;
-import org.wso2.carbon.user.core.UserStoreManager;
-import org.wso2.carbon.user.core.util.UserCoreUtil;
-import org.wso2.carbon.utils.multitenancy.MultitenantUtils;
-
-import java.io.IOException;
-import java.net.URLEncoder;
-import java.util.HashMap;
-import java.util.Map;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/**
- * Identifier based Authenticator
- */
-public class IdentifierAuthenticator extends AbstractApplicationAuthenticator
-        implements LocalApplicationExecutor {
-
-    private static final long serialVersionUID = 1819664539416029785L;
-    private static final Log log = LogFactory.getLog(IdentifierAuthenticator.class);
-    private static String RE_CAPTCHA_USER_DOMAIN = "user-domain-recaptcha";
-
-    @Override
-    public boolean canHandle(HttpServletRequest request) {
-        String userName = request.getParameter(IdentifierAuthenticatorConstants.USER_NAME);
-        return userName != null;
-    }
-
-    @Override
-    public AuthenticatorFlowStatus process(HttpServletRequest request,
-                                           HttpServletResponse response, AuthenticationContext context)
-            throws AuthenticationFailedException, LogoutFailedException {
-
-        if (context.isLogoutRequest()) {
-            return AuthenticatorFlowStatus.SUCCESS_COMPLETED;
-        } else {
-            return super.process(request, response, context);
-        }
-    }
-
-    @Override
-    protected void initiateAuthenticationRequest(HttpServletRequest request,
-                                                 HttpServletResponse response, AuthenticationContext context)
-            throws AuthenticationFailedException {
-
-        Map<String, String> parameterMap = getAuthenticatorConfig().getParameterMap();
-        String showAuthFailureReason = null;
-        if (parameterMap != null) {
-            showAuthFailureReason = parameterMap.get("showAuthFailureReason");
-            if (log.isDebugEnabled()) {
-                log.debug("showAuthFailureReason has been set as : " + showAuthFailureReason);
-            }
-        }
-
-        String loginPage = ConfigurationFacade.getInstance().getAuthenticationEndpointURL();
-        String retryPage = ConfigurationFacade.getInstance().getAuthenticationEndpointRetryURL();
-        String queryParams = context.getContextIdIncludedQueryParams();
-
-        try {
-            String retryParam = "";
-
-            if (context.isRetrying()) {
-                retryParam = "&authFailure=true&authFailureMsg=username.fail.message";
-            }
-
-            if (context.getProperty("UserTenantDomainMismatch") != null &&
-                    (Boolean) context.getProperty("UserTenantDomainMismatch")) {
-                retryParam = "&authFailure=true&authFailureMsg=user.tenant.domain.mismatch.message";
-                context.setProperty("UserTenantDomainMismatch", false);
-            }
-
-            IdentityErrorMsgContext errorContext = IdentityUtil.getIdentityErrorMsg();
-            IdentityUtil.clearIdentityErrorMsg();
-
-            if (errorContext != null && errorContext.getErrorCode() != null) {
-                if (log.isDebugEnabled()) {
-                    log.debug("Identity error message context is not null");
-                }
-                String errorCode = errorContext.getErrorCode();
-
-                if (errorCode.equals(IdentityCoreConstants.USER_ACCOUNT_NOT_CONFIRMED_ERROR_CODE)) {
-                    retryParam = "&authFailure=true&authFailureMsg=account.confirmation.pending";
-                    String username = request.getParameter(IdentifierAuthenticatorConstants.USER_NAME);
-                    Object domain = IdentityUtil.threadLocalProperties.get().get(RE_CAPTCHA_USER_DOMAIN);
-                    if (domain != null) {
-                        username = IdentityUtil.addDomainToName(username, domain.toString());
-                    }
-
-                    String redirectURL = loginPage + ("?" + queryParams) + IdentifierAuthenticatorConstants.FAILED_USERNAME
-                            + URLEncoder.encode(username, IdentifierAuthenticatorConstants.UTF_8) +
-                            IdentifierAuthenticatorConstants.ERROR_CODE + errorCode + IdentifierAuthenticatorConstants
-                            .AUTHENTICATORS + getName() + ":" + IdentifierAuthenticatorConstants.LOCAL + retryParam;
-                    response.sendRedirect(redirectURL);
-
-                } else if (showAuthFailureReason != null && "true".equals(showAuthFailureReason)) {
-
-                    String reason = null;
-                    if (errorCode.contains(":")) {
-                        String[] errorCodeReason = errorCode.split(":");
-                        errorCode = errorCodeReason[0];
-                        if (errorCodeReason.length > 1) {
-                            reason = errorCodeReason[1];
-                        }
-                    }
-                    int remainingAttempts =
-                            errorContext.getMaximumLoginAttempts() - errorContext.getFailedLoginAttempts();
-
-                    if (log.isDebugEnabled()) {
-                        log.debug("errorCode : " + errorCode);
-                        log.debug("username : " + request.getParameter(IdentifierAuthenticatorConstants.USER_NAME));
-                        log.debug("remainingAttempts : " + remainingAttempts);
-                    }
-
-                    if (errorCode.equals(UserCoreConstants.ErrorCode.INVALID_CREDENTIAL)) {
-                        retryParam = retryParam + IdentifierAuthenticatorConstants.ERROR_CODE + errorCode
-                                + IdentifierAuthenticatorConstants.FAILED_USERNAME + URLEncoder
-                                .encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                        IdentifierAuthenticatorConstants.UTF_8)
-                                + "&remainingAttempts=" + remainingAttempts;
-                        response.sendRedirect(loginPage + ("?" + queryParams)
-                                + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":" +
-                                IdentifierAuthenticatorConstants.LOCAL + retryParam);
-                    } else if (errorCode.equals(UserCoreConstants.ErrorCode.USER_IS_LOCKED)) {
-                        String redirectURL = retryPage;
-                        if (remainingAttempts == 0) {
-                            if (StringUtils.isBlank(reason)) {
-                                redirectURL = response.encodeRedirectURL(redirectURL + ("?" + queryParams)) +
-                                        IdentifierAuthenticatorConstants.ERROR_CODE + errorCode + IdentifierAuthenticatorConstants.FAILED_USERNAME +
-                                        URLEncoder.encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME), IdentifierAuthenticatorConstants.UTF_8) +
-                                        "&remainingAttempts=0";
-                            } else {
-                                redirectURL = response.encodeRedirectURL(redirectURL + ("?" + queryParams)) +
-                                        IdentifierAuthenticatorConstants.ERROR_CODE + errorCode + "&lockedReason="
-                                        + reason + IdentifierAuthenticatorConstants.FAILED_USERNAME +
-                                        URLEncoder.encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                                IdentifierAuthenticatorConstants.UTF_8) + "&remainingAttempts=0";
-                            }
-                        } else {
-                            if (StringUtils.isBlank(reason)) {
-                                redirectURL = response.encodeRedirectURL(redirectURL + ("?" + queryParams)) +
-                                        IdentifierAuthenticatorConstants.ERROR_CODE + errorCode + IdentifierAuthenticatorConstants.FAILED_USERNAME +
-                                        URLEncoder.encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME), IdentifierAuthenticatorConstants.UTF_8);
-                            } else {
-                                redirectURL = response.encodeRedirectURL(redirectURL + ("?" + queryParams)) +
-                                        IdentifierAuthenticatorConstants.ERROR_CODE + errorCode + "&lockedReason="
-                                        + reason + IdentifierAuthenticatorConstants.FAILED_USERNAME +
-                                        URLEncoder.encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                                IdentifierAuthenticatorConstants.UTF_8);
-                            }
-                        }
-                        response.sendRedirect(redirectURL);
-
-                    } else if (errorCode.equals(UserCoreConstants.ErrorCode.USER_DOES_NOT_EXIST)) {
-                        retryParam = retryParam + IdentifierAuthenticatorConstants.ERROR_CODE + errorCode
-                                + IdentifierAuthenticatorConstants.FAILED_USERNAME + URLEncoder
-                                .encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                        IdentifierAuthenticatorConstants.UTF_8);
-                        response.sendRedirect(loginPage + ("?" + queryParams)
-                                + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":" +
-                                IdentifierAuthenticatorConstants.LOCAL + retryParam);
-                    } else if (errorCode.equals(IdentityCoreConstants.USER_ACCOUNT_DISABLED_ERROR_CODE)) {
-                        retryParam = retryParam + IdentifierAuthenticatorConstants.ERROR_CODE + errorCode
-                                + IdentifierAuthenticatorConstants.FAILED_USERNAME + URLEncoder
-                                .encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                        IdentifierAuthenticatorConstants.UTF_8);
-                        response.sendRedirect(loginPage + ("?" + queryParams)
-                                + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":" +
-                                IdentifierAuthenticatorConstants.LOCAL + retryParam);
-                    } else {
-                        retryParam = retryParam + IdentifierAuthenticatorConstants.ERROR_CODE + errorCode
-                                + IdentifierAuthenticatorConstants.FAILED_USERNAME + URLEncoder
-                                .encode(request.getParameter(IdentifierAuthenticatorConstants.USER_NAME),
-                                        IdentifierAuthenticatorConstants.UTF_8);
-                        response.sendRedirect(loginPage + ("?" + queryParams)
-                                + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":"
-                                + IdentifierAuthenticatorConstants.LOCAL + retryParam);
-                    }
-                } else {
-                    if (log.isDebugEnabled()) {
-                        log.debug("Unknown identity error code.");
-                    }
-                    response.sendRedirect(loginPage + ("?" + queryParams)
-                            + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":" +
-                            IdentifierAuthenticatorConstants.LOCAL + retryParam);
-
-                }
-            } else {
-                if (log.isDebugEnabled()) {
-                    log.debug("Identity error message context is null");
-                }
-                response.sendRedirect(loginPage + ("?" + queryParams)
-                        + IdentifierAuthenticatorConstants.AUTHENTICATORS + getName() + ":" +
-                        IdentifierAuthenticatorConstants.LOCAL + retryParam);
-            }
-
-
-        } catch (IOException e) {
-            throw new AuthenticationFailedException(e.getMessage(), User.getUserFromUserName(request.getParameter
-                    (IdentifierAuthenticatorConstants.USER_NAME)), e);
-        }
-    }
-
-
-    @Override
-    protected void processAuthenticationResponse(HttpServletRequest request,
-                                                 HttpServletResponse response, AuthenticationContext context)
-            throws AuthenticationFailedException {
-
-        String username = request.getParameter(IdentifierAuthenticatorConstants.USER_NAME);
-
-        Map<String, Object> authProperties = context.getProperties();
-        if (authProperties == null) {
-            authProperties = new HashMap<String, Object>();
-            context.setProperties(authProperties);
-        }
-
-        boolean isAuthenticated;
-        UserStoreManager userStoreManager;
-        // Check the authentication
-        try {
-            int tenantId = IdentityTenantUtil.getTenantIdOfUser(username);
-            UserRealm userRealm = IdentifierAuthenticatorServiceComponent.getRealmService().getTenantUserRealm(tenantId);
-            if (userRealm != null) {
-                userStoreManager = (UserStoreManager) userRealm.getUserStoreManager();
-                isAuthenticated = userStoreManager.isExistingUser(MultitenantUtils.getTenantAwareUsername(username));
-            } else {
-                throw new AuthenticationFailedException("Cannot find the user realm for the given tenant: " +
-                        tenantId, User.getUserFromUserName(username));
-            }
-        } catch (IdentityRuntimeException e) {
-            if (log.isDebugEnabled()) {
-                log.debug("BasicAuthentication failed while trying to get the tenant ID of the user " + username, e);
-            }
-            throw new AuthenticationFailedException(e.getMessage(), User.getUserFromUserName(username), e);
-        } catch (org.wso2.carbon.user.api.UserStoreException e) {
-            if (log.isDebugEnabled()) {
-                log.debug("BasicAuthentication failed while trying to authenticate", e);
-            }
-            throw new AuthenticationFailedException(e.getMessage(), User.getUserFromUserName(username), e);
-        }
-
-        if (!isAuthenticated) {
-            if (log.isDebugEnabled()) {
-                log.debug("User authentication failed due to invalid credentials");
-            }
-            if (IdentityUtil.threadLocalProperties.get().get(RE_CAPTCHA_USER_DOMAIN) != null) {
-                username = IdentityUtil.addDomainToName(
-                        username, IdentityUtil.threadLocalProperties.get().get(RE_CAPTCHA_USER_DOMAIN).toString());
-            }
-            IdentityUtil.threadLocalProperties.get().remove(RE_CAPTCHA_USER_DOMAIN);
-            throw new InvalidCredentialsException("User authentication failed due to invalid credentials",
-                    User.getUserFromUserName(username));
-        }
-
-
-        String tenantDomain = MultitenantUtils.getTenantDomain(username);
-
-        //TODO: user tenant domain has to be an attribute in the AuthenticationContext
-        authProperties.put("user-tenant-domain", tenantDomain);
-
-        username = FrameworkUtils.prependUserStoreDomainToName(username);
-        authProperties.put("username", username);
-
-        if (getAuthenticatorConfig().getParameterMap() != null) {
-            String userNameUri = getAuthenticatorConfig().getParameterMap().get("UserNameAttributeClaimUri");
-            if (StringUtils.isNotBlank(userNameUri)) {
-                boolean multipleAttributeEnable;
-                String domain = UserCoreUtil.getDomainFromThreadLocal();
-                if (StringUtils.isNotBlank(domain)) {
-                    multipleAttributeEnable = Boolean.parseBoolean(userStoreManager.getSecondaryUserStoreManager(domain)
-                            .getRealmConfiguration().getUserStoreProperty("MultipleAttributeEnable"));
-                } else {
-                    multipleAttributeEnable = Boolean.parseBoolean(userStoreManager.
-                            getRealmConfiguration().getUserStoreProperty("MultipleAttributeEnable"));
-                }
-                if (multipleAttributeEnable) {
-                    try {
-                        if (log.isDebugEnabled()) {
-                            log.debug("Searching for UserNameAttribute value for user " + username +
-                                    " for claim uri : " + userNameUri);
-                        }
-                        String usernameValue = userStoreManager.
-                                getUserClaimValue(MultitenantUtils.getTenantAwareUsername(username), userNameUri, null);
-                        if (StringUtils.isNotBlank(usernameValue)) {
-                            tenantDomain = MultitenantUtils.getTenantDomain(username);
-                            usernameValue = FrameworkUtils.prependUserStoreDomainToName(usernameValue);
-                            username = usernameValue + "@" + tenantDomain;
-                            if (log.isDebugEnabled()) {
-                                log.debug("UserNameAttribute is found for user. Value is :  " + username);
-                            }
-                        }
-                    } catch (UserStoreException e) {
-                        //ignore  but log in debug
-                        if (log.isDebugEnabled()) {
-                            log.debug("Error while retrieving UserNameAttribute for user : " + username, e);
-                        }
-                    }
-                } else {
-                    if (log.isDebugEnabled()) {
-                        log.debug("MultipleAttribute is not enabled for user store domain : " + domain + " " +
-                                "Therefore UserNameAttribute is not retrieved");
-                    }
-                }
-            }
-        }
-        context.setSubject(AuthenticatedUser.createLocalAuthenticatedUserFromSubjectIdentifier(username));
-    }
-
-    @Override
-    protected boolean retryAuthenticationEnabled() {
-        return true;
-    }
-
-    @Override
-    public String getContextIdentifier(HttpServletRequest request) {
-        return request.getParameter("sessionDataKey");
-    }
-
-    @Override
-    public String getFriendlyName() {
-        return IdentifierAuthenticatorConstants.AUTHENTICATOR_FRIENDLY_NAME;
-    }
-
-    @Override
-    public String getName() {
-        return IdentifierAuthenticatorConstants.AUTHENTICATOR_NAME;
-    }
-}
