+++ pom.xml
@@ -45,7 +45,7 @@
 
     <modules>
         <module>components/org.wso2.carbon.identity.application.authenticator.basicauth</module>
-        <module>components/org.wso2.carbon.identity.application.authenticator.identifierauth</module>
+        <module>components/org.wso2.carbon.identity.application.authentication.handler.identifier</module>
         <module>features/org.wso2.carbon.identity.application.authenticator.basicauth.server.feature</module>
         <module>components/org.wso2.carbon.identity.application.authenticator.basicauth.jwt</module>
     </modules>
@@ -372,7 +372,7 @@
         <carbon.base.imp.pkg.version.range>[1.0.0, 2.0.0)</carbon.base.imp.pkg.version.range>
 
         <!-- Carbon Identity Framework version -->
-        <carbon.identity.framework.version>5.12.56</carbon.identity.framework.version>
+        <carbon.identity.framework.version>5.12.113</carbon.identity.framework.version>
         <carbon.identity.framework.imp.pkg.version.range>[5.7.0, 6.0.0)
         </carbon.identity.framework.imp.pkg.version.range>
